23.10.15 2:03

+++ Task 04/4-04/Program.cs
@@ -7,7 +7,7 @@
     {
         private static void Main(string[] args)
         {
-            int[] array = new int[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+            byte[] array = new byte[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
 
             Console.WriteLine(array.ArraySum());
         }
